Move setup_logging into CSET._common
We are going to use it in multiple places now.
