+++ tests/test_cli.py
@@ -110,30 +110,6 @@ def test_argument_parser_extract_workflow(tmp_path):
     assert args.location == tmp_path
 
 
-def test_setup_logging():
-    """Tests the logging setup at various verbosity levels."""
-    root_logger = logging.getLogger()
-    # Log has a minimum of WARNING.
-    CSET.setup_logging(0)
-    assert root_logger.level == logging.WARNING
-    # -v
-    CSET.setup_logging(1)
-    assert root_logger.level == logging.INFO
-    # -vv
-    CSET.setup_logging(2)
-    assert root_logger.level == logging.DEBUG
-
-
-def test_setup_logging_mpl_font_logs_filtered(caplog):
-    """Test matplotlib log messages about fonts are filtered out."""
-    CSET.setup_logging(2)
-    logger = logging.getLogger("matplotlib.font_manager")
-    logger.debug("findfont: message")
-    logger.debug("other message")
-    assert len(caplog.records) == 1
-    assert caplog.records[0].getMessage() == "other message"
-
-
 def test_main_no_subparser(capsys):
     """Appropriate error when no subparser is given."""
     with pytest.raises(SystemExit) as sysexit:
